+++ src/Phase2Task/Phase2Task/ExpenseTracker/ExpenseTracker/Data/Expense.json
@@ -1,12 +1,1 @@
-[
-  {
-    "EntryDate": "2023-12-08T00:00:00+05:30",
-    "Amount": 400,
-    "Category": "Tea"
-  },
-  {
-    "EntryDate": "2023-12-08T00:00:00+05:30",
-    "Amount": 200,
-    "Category": "Tea"
-  }
-]
+[]
